# server/resource/Cleverbuys_db/User.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE User.js PLEASE EDIT ../custom/UserCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Cleverbuys_db = require('../../db/Cleverbuys_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/UserCustom.js');

/*
 * SCHEMA DB User
 * 
	{
		lat: {
			type: 'Number'
		},
		lng: {
			type: 'Number'
		},
		mail: {
			type: 'String'
		},
		name: {
			type: 'String'
		},
		password: {
			type: 'String', 
			required : true
		},
		postCode: {
			type: 'String'
		},
		roles: {
			type: 'String'
		},
		state: {
			type: 'String'
		},
		surname: {
			type: 'String'
		},
		town: {
			type: 'String'
		},
		username: {
			type: 'String', 
			required : true
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		
	}
 * 
 */



//CRUD METHODS


/**
 * UserService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/user', function (req, res) {
	obj = new db_Cleverbuys_db.User(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * UserService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id
 *
 */
app['delete'](properties.api + '/user/:id', function (req, res) {
	db_Cleverbuys_db.User.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * UserService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id resource
 *
 */
app.get(properties.api + '/user/:id', function (req, res) {
	db_Cleverbuys_db.User.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * UserService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/user', function (req, res) {
	db_Cleverbuys_db.User.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * UserService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id
 *
 */
app.post(properties.api + '/user/:id', function (req, res) {
	db_Cleverbuys_db.User.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */


/**
 * UserService.changePassword
 *   @description Change password of user from admin
 *   @returns object
 *
 */
app['post'](properties.api + '/user/:id/changePassword', function (req, res) {
	res.send({});
});
		