# server/swagger.yaml
openapi: 3.0.0
info:
  title: Cleverbuys
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user/:
    post:
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                postCode:
                  type: string
                roles:
                  type: string
                state:
                  type: string
                surname:
                  type: string
                town:
                  type: string
                username:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  brands  -----
  /brands/:
    post:
      summary: CRUD ACTION create
      tags:
        - brands
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/brands"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - brands
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/brands"
  /brands/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - brands
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - brands
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/brands"
    post:
      summary: CRUD ACTION update
      tags:
        - brands
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/brands"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/brands"
    
  #-----  dealerBrands  -----
  /dealerbrands/:
    post:
      summary: CRUD ACTION create
      tags:
        - dealerBrands
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dealerBrands"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                brandID:
                  type: number
                dealershipID:
                  type: number
    get:
      summary: CRUD ACTION list
      tags:
        - dealerBrands
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dealerBrands"
  /dealerbrands/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - dealerBrands
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - dealerBrands
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dealerBrands"
    post:
      summary: CRUD ACTION update
      tags:
        - dealerBrands
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dealerBrands"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dealerBrands"
  /dealerbrands/findBybrandID/{key}:
    get:
      summary: CRUD ACTION findBybrandID
      tags:
        - dealerBrands
      parameters:
        - name: key
          in: path
          description: Id of the resource brandID to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dealerBrands"
    
  #-----  dealership  -----
...
  /Users/{id}/changePassword:
    post:
      summary: Change password by admin API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordAdmin:
                  type: string
      responses:
        '200':
          description: OK
  /Users/{id}:
    post:
      summary: Update User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    delete:
      summary: Delete User API
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      properties:
        _id:
          type: string
        lat:
          type: number
        lng:
          type: number
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        postCode:
          type: string
        roles:
          type: string
        state:
          type: string
        surname:
          type: string
        town:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username
    brands:
      properties:
        _id:
          type: string
        name:
          type: string
      required:
        -  _id
        - name
    dealerBrands:
      properties:
        _id:
          type: string
        brandID:
          type: number
        dealershipID:
          type: number
      required:
        -  _id
        - brandID
        - dealershipID
...